pyproject.toml
[project]
name = "omopy"
version = "0.1.0"
description = "Pythonic, type-safe interface for OMOP CDM databases"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.14"
authors = [{ name = "DARWIN EU Coordination Centre" }]
keywords = [
    "omop",
    "cdm",
    "ohdsi",
    "observational-health",
    "epidemiology",
    "fastomop",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Typing :: Typed",
]

dependencies = [
    "duckdb>=1.5.0",
    "ibis-framework>=12.0.0",
    "pandas>=2.3.3",
    "polars>=1.0",
    "pydantic>=2.12",
    "pyarrow>=23.0.1",
    "pyarrow-hotfix>=0.7",
    "sqlalchemy>=2.0.48",
    "sqlalchemy-utils>=0.42.1",
    "great-tables>=0.21.0",
    "plotly>=6.6.0",
    "scipy>=1.15.0",
    "lifelines>=0.29",
    "openpyxl>=3.1.5",
    "nbconvert>=7.0",
]

[project.urls]
Homepage = "https://github.com/fastomop/omopy"
Documentation = "https://fastomop.github.io/omopy"
Repository = "https://github.com/fastomop/omopy"
Issues = "https://github.com/fastomop/omopy/issues"
Changelog = "https://github.com/fastomop/omopy/releases"

[project.optional-dependencies]
# Database backends (Phase 1+)
postgres = [
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.1",
    "ibis-framework[postgres]>=9.0",
]
snowflake = [
    "sqlalchemy>=2.0",
    "snowflake-sqlalchemy>=1.5",
    "ibis-framework[snowflake]>=9.0",
]
bigquery = [
    "sqlalchemy>=2.0",
    "sqlalchemy-bigquery>=1.8",
    "ibis-framework[bigquery]>=9.0",
]
mssql = ["sqlalchemy>=2.0", "pyodbc>=5.0", "ibis-framework[mssql]>=9.0"]
spark = ["sqlalchemy>=2.0", "pyspark>=3.5", "ibis-framework[pyspark]>=9.0"]
oracle = ["sqlalchemy>=2.0", "oracledb>=2.0", "ibis-framework[oracle]>=9.0"]
duckdb = ["sqlalchemy>=2.0", "duckdb>=1.0", "ibis-framework[duckdb]>=9.0"]

# Analytical extras (Phase 2+)
vis = ["plotnine>=0.13", "great-tables>=0.10"]
survival = ["lifelines>=0.29"]

# Everything
all = [
    "omopy[duckdb,postgres,mssql,snowflake,bigquery,spark,oracle,vis,survival]",
]

[dependency-groups]
dev = [
    "hypothesis>=6.100",
    "jupyterlab>=4.5.6",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.6",
    "mypy>=1.10",
    "pre-commit>=3.7",
    "prek>=0.3.8",
    "pytest-cov>=5.0",
    "pytest>=8.0",
    "ruff>=0.5",
    "syrupy>=4.6",
    "mkdocstrings>=1.0.3",
    "mkdocstrings-python>=2.0.3",
]
docs = [
    "mkdocs-material>=9.7.6",
    "mkdocstrings>=1.0.3",
    "mkdocstrings-python>=2.0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/omopy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["-ra", "--strict-markers", "--strict-config"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests requiring database connections",
]

[tool.ruff]
target-version = "py314"
line-length = 88
indent-width = 4
exclude = ["build", "dist", "omopy.egg-info", ".venv"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "RUF"]
ignore = [
    "A001",    # variable shadowing builtin (e.g. `type` module-level)
    "A002",    # argument shadowing builtin — `type` param is an API convention
    "RUF002",  # ambiguous Unicode in docstrings (×, –) — intentional
    "RUF003",  # ambiguous Unicode in comments (×, –) — intentional
    "RUF022",  # `__all__` not sorted — we group by category for readability
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B017"]  # `assertRaises(Exception)` OK in tests

[tool.mypy]
python_version = "3.14"
strict = true
warn_unused_ignores = true
